♻️ refactor: Refactor stream highlighter
